update starting stat thalia ent

+++ app/core/include/common.h
@@ -11,15 +11,15 @@ struct Stats{
     int m_ArmPhy = 0;
     int m_ArmMag = 0;
     int m_powPhy = 0;
-    int m_powMag = 0;
+    double m_powMag = 0;
     int m_aggro = 0;
     int m_speed = 0;
     // critical strike in %
     int m_CriticalStrike  = 0;
     // dodge in %
-    float m_dogde = 0;
+    double m_dogde = 0;
     int m_regenHP = 0;
-    int m_regenMana = 0;
+    double m_regenMana = 0;
     int m_regenVigueur = 0;
 };
